ec2-instance-selector: static bottles (#108)
* ec2-instance-selector: use static bottles

* fix building static bottles

* add indv build-formula success to its own file rather than relying on the bottle output

+++ build-formula.sh
@@ -5,8 +5,9 @@ set -euo pipefail
 # This allows it to be executed from anywhere on the file system.
 SCRIPTPATH="$( cd "$(dirname "$0")" ; pwd -P )"
 BUILD_DIR="$SCRIPTPATH/build"
+RESULTS_DIR="$BUILD_DIR/results"
 # Create temp dir at root to hold files during this execution
-mkdir -p $BUILD_DIR
+mkdir -p $BUILD_DIR $RESULTS_DIR
 
 export HOMEBREW_NO_INSTALL_CLEANUP=1
 # Commenting out the auto upgrade since appveyor mac has an old version and fails buildilng bottles for SAM CLI
@@ -90,9 +91,9 @@ function fail_msg() {
     exit 2
 }
 
-trap fail_msg err int term kill
+trap fail_msg err int term
 
-brew untap ${SAVED_TAPS[@]} 2>/dev/null || :
+brew untap "${SAVED_TAPS[@]}" 2>/dev/null || :
 
 if [[ ${USER_TAP} != "" ]]; then
   TAP="${USER_TAP}"
@@ -132,22 +133,24 @@ echo "[${BOTTLE}]: Root URL -> ${BOTTLE_ASSET_URL}"
 BIN_NAME=$(echo ${BOTTLE_CONFIG} | jq -r '.bin')
 echo "[${BOTTLE}]: Bottle Binary -> ${BIN_NAME}"
 
-# Need to cd since brew bottle doesn't have an option to 
-cd "${BUILD_DIR}"
-brew bottle --no-rebuild --json --root-url="${BOTTLE_ASSET_URL}" "${TAP}/${BOTTLE}"
-cd ${SCRIPTPATH}
+if [[ $(grep -c 'bottle :unneeded' "${FORMULA_FILE}") -eq 0 ]]; then 
+  # Need to cd since brew bottle doesn't have an option to 
+  cd "${BUILD_DIR}"
+  brew bottle --no-rebuild --json --root-url="${BOTTLE_ASSET_URL}" "${TAP}/${BOTTLE}"
+  cd ${SCRIPTPATH}
 
-RELEASE_FILE="$(ls ${BUILD_DIR}/${BOTTLE}--*.bottle.tar.gz)"
-# Renaming aws-sam-cli--0.37.0.sierra.bottle.tar.gz to aws-sam-cli-0.37.0.sierra.bottle.tar.gz
-rename 's/--/-/' ${BUILD_DIR}/*.bottle.* # replacing `--` with `-`
-RELEASE_FILE=$(echo ${RELEASE_FILE} | sed 's/--/-/')
+  RELEASE_FILE="$(ls ${BUILD_DIR}/${BOTTLE}--*.bottle.tar.gz)"
+  # Renaming aws-sam-cli--0.37.0.sierra.bottle.tar.gz to aws-sam-cli-0.37.0.sierra.bottle.tar.gz
+  rename 's/--/-/' ${BUILD_DIR}/*.bottle.* # replacing `--` with `-`
+  RELEASE_FILE=$(echo ${RELEASE_FILE} | sed 's/--/-/')
 
-echo "[${BOTTLE}]: Brew bottles are built. Validating them now."
-echo "[${BOTTLE}]: Release file -> ${RELEASE_FILE}"
+  echo "[${BOTTLE}]: Brew bottles are built. Validating them now."
+  echo "[${BOTTLE}]: Release file -> ${RELEASE_FILE}"
 
-brew uninstall -f ${BOTTLE}
-echo "[${BOTTLE}]: Installing release file"
-brew install ${RELEASE_FILE}
+  brew uninstall -f ${BOTTLE}
+  echo "[${BOTTLE}]: Installing release file"
+  brew install ${RELEASE_FILE}
+fi
 
 BUILT_BOTTLE_VERSION="$(${BIN_NAME} --version | grep -Eo '[0-9]+\.[0-9]+\.[0-9a-zA-Z]+')"
 if [[ "${BOTTLE_ASSET_VERSION}" != "${BUILT_BOTTLE_VERSION}" ]]; then
@@ -156,6 +159,8 @@ if [[ "${BOTTLE_ASSET_VERSION}" != "${BUILT_BOTTLE_VERSION}" ]]; then
 fi
 
 echo "✅ [${BOTTLE}]: Verified that the new ${BOTTLE} version ${BUILT_BOTTLE_VERSION} is the same as what is expected ${BOTTLE_ASSET_VERSION} 🎉🥂✅"
+## Touch results file for build-bottles summary
+touch "${RESULTS_DIR}/${BOTTLE}"
 
 brew uninstall -f ${BOTTLE}
